Aurhorization for edit and delete

+++ app/Http/Controllers/ListingController.php
@@ -64,6 +64,10 @@ class ListingController extends Controller
   // Update Listing Data
   public function update(Request $request, Listing $listing)
   {
+    if($listing->user_id !== auth()->id()) {
+      abort(403, 'Unauthorized Action');
+    }
+
     $formFields = $request->validate([
       'title' => 'required',
       'company' => 'required',
@@ -87,6 +91,9 @@ class ListingController extends Controller
 
   //Delete Listing
   public function delete(Listing $listing) {
+    if($listing->user_id !== auth()->id()) {
+      abort(403, 'Unauthorized Action');
+    }
     $listing->delete();
 
     return redirect('/')->with('message', 'Listing deleted successfully');
